Refactoring of serializers/*.py and fields.py
